--- frontend/src/pages/about-lisitaBSB/style.module.css
@@ -1,168 +1,0 @@
-.imagem1{
-    width: 15rem;
-    height: 15rem;
-    border-radius: 50%;
-    box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.3);
-    object-fit: cover;
-}
-
-.imageContext{
-    display: flex;
-    flex-direction: row;
-    gap: 1.5rem;
-    align-items: center;
-    justify-content: space-around;
-}
-
-.divImageContext{
-    width: 65%;
-}
-
-.paragraph1Context{
-    margin-left: 0;
-    padding: 0;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-.paragraph2Context{
-    margin-left: 0;
-    display: flex;
-    flex-direction: space-between;
-    justify-content: center;
-    align-items: center;
-}
-.mainContext{
-    margin-top: 3rem;
-    margin-left: 22.5%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 55%;
-}
-
-.cardAboutContext {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    background-color: #fff;
-    width: 100%;
-    padding-top: 32px;
-    padding-left: 40px;
-    padding-right: 40px;
-    padding-bottom: 32px;
-    box-shadow: 0 0 5px rgb(177, 175, 175);
-}
-
-.cardAboutTitle {
-    font-size: 40px;
-    margin-left: 0;
-    font-weight: normal;
-    text-align: center;
-}
-
-.cardAboutSubtitle {
-    margin-left: 0;
-    font-size: 32px;
-    margin: 1rem;
-    width: 100%;
-    font-weight: normal;
-    text-align: start;
-}
-
-.cardAboutSubtitle3{
-    width: 100%;
-    font-size: 20px;
-    margin-left: 0;
-    font-weight: normal;
-    text-align: start;
-}
-
-.cardAboutTextP {
-    width: 100%;
-    margin: .5rem;
-    margin-left: 0;
-    font-size: 15px;
-    text-align: justify;
-}
-
-@media(max-width: 1200px){
-    .mainContext{
-        margin-top: 3rem;
-        margin-left: 12.5%;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        width: 75%;
-    }
-    .cardAboutImg{
-        margin-top: 16px;
-        width: 20rem;
-    }
-
-    .cardAboutTitle {
-        font-size: 30px;
-        margin-left: 0;
-        font-weight: normal;
-        text-align: center;
-    }
-    
-    .cardAboutSubtitle {
-        margin-left: 0;
-        font-size: 24px;
-        margin: 1rem;
-        font-weight: normal;
-        text-align: start;
-    }
-    
-    .cardAboutSubtitle3{
-        font-size: 16px;
-        margin-left: 0;
-        font-weight: normal;
-        text-align: start;
-    }
-    
-    .cardAboutTextP {
-        margin: .3rem;
-        margin-left: 0;
-        font-size: 15px;
-        text-align: justify;
-    }
-}
-
-@media(max-width: 800px){
-    body{
-        background-color: #F9F9F9;
-    }
-    .mainContext{
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        width: 90%;
-        margin: 0;
-        margin-left: 5%;
-        margin-top: 1.5rem;
-        padding: 0;
-        background-color: #F9F9F9;
-    }
-    .cardAboutContext {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        justify-content: center;
-        width: 100%;
-        padding-top: 0px;
-        padding-left: 0px;
-        padding-right: 0px;
-        padding-bottom: 0px;
-        border: 0;
-        margin: 0;
-        box-shadow: 0 0 transparent;
-        background-color: #F9F9F9;
-        
-    }
-}
